+++ example/src/plugins/fade/FadeTilesPlugin.js
